refactored addPhoto and refactoring orderPhoto

+++ app/src/main/java/com/beeswork/balance/data/database/entity/Photo.kt
@@ -10,7 +10,7 @@ data class Photo(
     @PrimaryKey
     val key: String,
 
-    val sequence: Long,
+    val sequence: Int,
     var synced: Boolean,
 ) {
     companion object {
